+++ config/ember-try.js
@@ -70,5 +70,6 @@ module.exports = {
         devDependencies: {}
       }
     }
-  ]
+  ],
+  useYarn: true
 };