+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - bump esri-loader to 2.3 to default to JSAPI 4.7
 
+### Changed
+- use yarn
+
 ## 2.2.0
 
 ### Added